+++ src/server/scripts/EasternKingdoms/BlackrockMountain/BlackrockDepths/boss_eviscerator.cpp
@@ -52,9 +52,9 @@ public:
         void EnterCombat(Unit* /*who*/) override
         {
             _EnterCombat();
-            events.ScheduleEvent(SPELL_SHADOWBOLT_VOLLEY, 0.2 * TIMER_SHADOWBOLT_VOLLEY);
-            events.ScheduleEvent(SPELL_REND, 0.2 * TIMER_REND);
-            events.ScheduleEvent(SPELL_SHIELD, 0.2 * TIMER_SHIELD);
+            events.ScheduleEvent(SPELL_SHADOWBOLT_VOLLEY, 0.2 * (int)TIMER_SHADOWBOLT_VOLLEY);
+            events.ScheduleEvent(SPELL_REND, 0.2 * (int) TIMER_REND);
+            events.ScheduleEvent(SPELL_SHIELD, 0.2 * (int) TIMER_SHIELD);
         }
 
         void DamageTaken(Unit* /* doneBy */, uint32& /* damage */, DamageEffectType /* damagetype */, SpellSchoolMask damageSchoolMask) override
